Implement dialog panel to confirm actions

+++ components/DataTable.tsx
@@ -25,6 +25,7 @@ import {
   TableHeader,
   TableRow,
 } from "@/components/ui/table";
+import { useConfirm } from "@/hooks/use-confirm";
 import { Button } from "./ui/button";
 import { Input } from "@/components/ui/input";
 import { Trash } from "lucide-react";
@@ -46,6 +47,10 @@ export function DataTable<TData, TValue>({
   onDelete,
   disabled,
 }: DataTableProps<TData, TValue>) {
+  const [ConfirmationDialog, confirm] = useConfirm(
+    "Are you sure?",
+    "You are about to perform a bulk delete."
+  )
   // State for sorting, filtering, and row selection
   const [sorting, setSorting] = React.useState<SortingState>([]);
   const [columnFilters, setColumnFilters] = React.useState<ColumnFiltersState>(
@@ -73,6 +78,7 @@ export function DataTable<TData, TValue>({
 
   return (
     <div>
+      <ConfirmationDialog />
       {/* Filter input and conditional delete button */}
       <div className="flex items-center py-4">
         <Input
@@ -89,9 +95,13 @@ export function DataTable<TData, TValue>({
             size="sm"
             variant="outline"
             className="ml-auto font-normal text-xs"
-            onClick={() => {
+            onClick={async () => {
+              const ok = await confirm();
+
+              if (ok) {
               onDelete(table.getFilteredSelectedRowModel().rows);
               table.resetRowSelection();
+              }
             }}
           >
             <Trash className="size-4 mr-2" />
